backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.4.0
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.0 